Clean up syntax errors
