+++ lib/src/timezones/tzdb_stream_reader.dart
@@ -1,38 +1,85 @@
 import 'dart:typed_data';
 
 import 'package:time_machine2/src/time_machine_internal.dart';
-import 'package:time_machine2/src/timezones/tzdb_stream_writer.dart';
-import 'package:time_machine2/time_machine2.dart';
+
+// from: https://github.com/nodatime/nodatime/blob/master/src/NodaTime/TimeZones/IO/TzdbStreamFieldId.cs
+
+/// Enumeration of the fields which can occur in a TZDB stream file.
+/// This enables the file to be self-describing to a reasonable extent.
+enum TzdbStreamFieldId {
+  /// String pool. Format is: number of strings (WriteCount) followed by that many string values.
+  /// The indexes into the resultant list are used for other strings in the file, in some fields.
+  stringPool,
+
+  /// Repeated field of time zones. Format is: zone ID, then zone as written by DateTimeZoneWriter.
+  timeZone,
+
+  /// Single field giving the version of the TZDB source data. A string value which does *not* use the string pool.
+  tzdbVersion,
+
+  /// Single field giving the mapping of ID to canonical ID, as written by DateTimeZoneWriter.WriteDictionary.
+  tzdbIdMap,
+
+  /// Single field containing mapping data as written by WindowsZones.Write.
+  cldrSupplementalWindowsZones,
+
+  /// Single field giving the mapping of Windows StandardName to TZDB canonical ID,
+  /// for time zones where TimeZoneInfo.Id != TimeZoneInfo.StandardName,
+  /// as written by DateTimeZoneWriter.WriteDictionary.
+  windowsAdditionalStandardNameToIdMapping,
+
+  /// Single field providing all zone locations. The format is simply a count, and then that many copies of
+  /// TzdbZoneLocation data.
+  zoneLocations,
+
+  /// Single field providing all 'zone 1970' locations. The format is simply a count, and then that many copies of
+  /// TzdbZone1970Location data. This field was introduced in Noda Time 2.0.
+  zone1970Locations
+}
+
+enum DateTimeZoneType {
+  fixed,
+  precalculated,
+}
 
 /// Reads time zone data from a stream in nzd format.
-/// This class reads the format written by TzdbStreamWriter.
+/// This class reads the format written by TzdbStreamWriter. It follows a different
+/// approach than the other stream reader classes and reads all stream data at
+/// once in the constructor, then publishes the results as properties of the
+/// class.
 class TzdbStreamReader {
   static const int _expectedVersion = 0;
 
   final ByteData _input;
   int _currentOffset = 0;
   late List<String> _stringPool;
+  late Map<TzdbStreamFieldId, List<int>> fields;
 
-  TzdbStreamReader(this._input);
+  late String tzdbVersion;
+  late Map<String, DateTimeZone> timeZones;
+  late Map<String, String> aliases;
 
-  /// Reads a complete TZDB database from the stream
-  Map<String, DateTimeZone> read() {
+  TzdbStreamReader(this._input) {
     // Read and validate the version
     int version = _input.getUint8(_currentOffset++);
     if (version != _expectedVersion) {
       throw Exception('Expected version $_expectedVersion but got $version');
     }
 
-    Map<TzdbStreamFieldId, List<int>> fields = _readFields();
-
-    // First read the string pool, as we'll need it for everything else
+    // read fields and string pool
+    fields = _readFields();
     _stringPool = _readStringPool(fields[TzdbStreamFieldId.stringPool]!);
 
     // Now read the actual data
     var reader = DateTimeZoneReader(_input, _stringPool);
 
+    // Read version
+    var versionField = fields[TzdbStreamFieldId.tzdbVersion]!;
+    reader.currentOffset = versionField[0];
+    tzdbVersion = reader.readString();
+
     // Read all time zones
-    var timeZones = <String, DateTimeZone>{};
+    timeZones = <String, DateTimeZone>{};
     var timeZoneField = fields[TzdbStreamFieldId.timeZone];
     if (timeZoneField != null) {
       reader.currentOffset = timeZoneField[0];
@@ -42,21 +89,17 @@ class TzdbStreamReader {
       }
     }
 
-    // Read version
-    var versionField = fields[TzdbStreamFieldId.tzdbVersion]!;
-    reader.currentOffset = versionField[0];
-    String tzdbVersion = reader.readString();
-
     // Read zone aliases
     var aliasField = fields[TzdbStreamFieldId.tzdbIdMap]!;
     reader.currentOffset = aliasField[0];
-    var aliases = reader.readDictionary();
+    aliases = reader.readDictionary();
 
+    /*
     // Read Windows mappings
     var windowsZonesField =
         fields[TzdbStreamFieldId.cldrSupplementalWindowsZones]!;
     reader.currentOffset = windowsZonesField[0];
-    var windowsZones = WindowsZones.read(reader);
+    windowsZones = WindowsZones.read(reader);
 
     // Read zone locations if present
     List<TzdbZoneLocation>? zoneLocations;
@@ -77,8 +120,7 @@ class TzdbStreamReader {
       zone1970Locations = List<TzdbZone1970Location>.generate(
           count, (_) => TzdbZone1970Location.read(reader));
     }
-
-    return timeZones;
+    */
   }
 
   Map<TzdbStreamFieldId, List<int>> _readFields() {
